A demo html file for plugin usage.
Javascript is used to create the code to be executed from the actual
text of the example.  Therefore they will always match up.

The $.ajax call is redefined so that all data transfer takes place
on the page itself.  There is a tiny in-page database that gets
updated when you edit any of the fields.