Put the stop propagation back in :)
This time we are binding directly to the submit button to cancel
the blur event.

We do not need to attach to the cancel event because... hmm well maybe
we do, but only in the case where there is a cancel button but not a
submit button and you are expecting submit on blur. This isn't
anything that I am likely to do.